Move all of the real code into a separate crate
