+++ experimental/uefi/baremetal/Cargo.toml
@@ -6,18 +6,5 @@ edition = "2021"
 license = "Apache-2.0"
 
 [dependencies]
-anyhow = { version = "*", default-features = false }
-atomic_refcell = "*"
-getrandom = { version = "0.2", features = ["rdrand"] }
-lazy_static = { version = "*", features = ["spin_no_std"] }
-libm = "*"
-linked_list_allocator = { version = "*" }
-log = "*"
-rand_core = { version = "*", features = ["getrandom"] }
-runtime = { path = "../runtime", default-features = false, features = [
-  "rust-crypto"
-] }
-uart_16550 = "*"
+kernel = { path = "../kernel" }
 rust-hypervisor-firmware-subset = { path = "../../../third_party/rust-hypervisor-firmware-subset" }
-x86_64 = "*"
-ciborium-io = { version = "*", default-features = false }
